WT-7230 Initial set of CMake configuration helpers
Added an initial set of CMake configuration helper functions.
These are intended to help with defining the various WiredTiger
config options in CMake. This will allow the user to define
different types of configuration variables that can be propagated
across the build system and ultimately exported to wiredtiger
configuration headers.