Environment-based CMake configuration helpers
To auto-detect various settings and resources in the users
toolchain environment, created a set of configuration helpers to
establish the existence of various symbols, include headers and
types. This will help bridge a subset of the host environment
detection used by the auto-conf build system for configuring our
wiredtiger build.