WT-7266 Test re-reading closed active hs files
Added a new test to validate we don't lose any data when idle
files with an active history are closed/sweeped. With WT-6819
files with active history, ie content newer than the oldest
timestamp can be closed when idle. This test ensures that when an
active history file is idle closed we can continue reading the
correct version of data.