WT-7373 Investigate slow random cursor operations on oplog (#6466)
If the tree is entirely populated with large items (MongoDB configures the maximum leaf page item to avoid overflow objects), we can still end up reading huge amounts of data on every random cursor sample. If the page is clean (that is, just read from disk), it can't be fully populated with deleted items, use it as our sample.