WT-7426 Set write generation number when the page image gets created (#6474)
Set write generation number whenever the page image is created instead of
waiting for it to be written to disk. In scenarios, when the page image is instantiated
it can lead to misuse of the stale transaction ids present on the page.

(cherry picked from commit b3e3b2580ce14e7609e7ba269d3c7432f81339b7)