WT-7460 RTS to abort all the updates from prepared transaction
It is only possible for the preapred transactions to have an on-disk
data with both start and stop time windows are same and their prepared
timestamp is less than the stable timestamp. RTS must abort all the
updates from the prepared transaction either by replace it with stable
data or remove the key.