WT-7460 RTS to abort all the updates from a prepared transaction (#6496)
It is only possible for the prepared transactions to have an on-disk
data with both start and stop time windows are the same and the
prepared timestamp is less than the stable timestamp. RTS must
abort all the updates from the prepared transaction either by
replacing it with stable data or remove the key.

(cherry picked from commit cf79d062ecdaa7509ca80c4a643268ee5303e915)