Assert that WT_CONNECTION.rollback_to_stable returns success.
When repeating reads after rollback-to-stable, don't bother checking if the repeatable read
timestamp is less than the global timestamp, the underlying read function handles that as an
EINVAL return from setting the read timestsamp.

When repeating reads (both in the normal case and after rollback-to-stable), handle WT_ROLLBACK
returns by ignoring the error and repeating the operation. Eviction may choose to rollback our
transaction, but it should eventually complete.

When repeating reads (both in the normal case and after rollback-to-stable), don't treat cache-full
as an expected error, these are read-only transactions which should not trigger cache-full checks.