Remove smoke.sh usage for POSIX+Windows tests
Given that shell scripts aren't portable for Windows environments,
replaced usage of smoke.sh with ctest variant definitions. This
being equivalent to the content in the smoke.sh scripts with the
added benefits of being portable and parallelizable. These changes
have only been applied to C tests that are expected to run on both
Windows and POSIX environments (the other smoke scripts can be left
to a future task).