Expanded testutil options with build dir option
Given the CMake build directory location is more dynamic than the
autoconf one e.g out-of-source build directory can be called
anything, its not longer safe to assume the build will exist in
'build_posix'. Expanded the testutil options so that users can
pass in a '-b' option to specify the build directory location.
This allowing build directory discovery to be used with CMake.