Manually specify build dir and WT_FAIL_FS_LIB
The c tests 'wt2909_checkpoint_integrity' & 'wt3120_filesys'
currently assume the location of the fail fs extension
library/module (as if we are in the build_posix directory). This
won't work with CMake however due to a different layout. Updated
the tests so we can dynamically specify the build directory and
fail fs extension location.