'wt4333_handle_locks' accept custom home directory
The CMake CTest definitions attempt to invoke the csuite tests
with a '-h <home path>' option to ensure we create unique
home DB paths. We do this so tests can be run in parallel and
the path will work within the new build directory structure.
Expanded the wt4333_handle_locks test to accept the '-h' option as
this currently doesn't use the testutil option parsing helpers.