Fixed issue with 'config_bool' defaults
When a dependency is not met under 'config_bool' is sets the
configs value to the specified default value. This causes issues
when it defaults to 'True' when dependencies are not met e.g
'WT_POSIX' would result to being set to 'True' on Windows
platforms due to failing the dependency check on
'WT_LINUX OR WT_DARWIN'. Fixed the logic to default to 'False' if
we fail a dependency check.