Build mode support on Windows
Added the ability to enable ASan and UBSan on Windows. ASan is now
support starting from Visual Studio 2019 version 16.9. This
specific version isn't exported to a CMake variable if not using
a VS-based generator. So refactored the mode config script to
perform compiler flag checks to see if the users VS compiler
supports ASan.