Opportunity to define WT_BUILDDIR in python test
When running the python test suite with a CMake generated build
we can no longer depend on falling back to a build directory
of 'build_posix'. Users can dynamically define the name of
their CMake build directory and create multiple working builds of
varying names. To avoid incorrectly assuming the build exists in
'build_posix', provide an opportunity to read an env variable
'WT_BUILDDIR' indicating the location of the build we want to use
for the test.