Loosen the restrictions around evicting fast-truncate pages to avoid cache-stuck failures.
Add a WT_PAGE_DELETED.resolved field that's non-zero once the truncate has been committed or
aborted (but says nothing about which resolution happened or data visibility).

When instantiating a fast-truncate page, use the WT_PAGE_DELETED.resolved field to determine
if we need to create the WT_PAGE_DELETED.update_list array of update references. This doesn't
directly help with evicting fast-truncate pages, but it avoids allocating an additional 8B of
memory for every page entry.

If failing to instantiate a fast-truncate page, don't increment the page's memory footprint.

Don't require global visibility before attempting to evict fast-truncate pages, instead permit
eviction to be attempted as soon as the fast-truncate resolves. This allows fully instantiated
pages to be reconciled and evicted and their memory reclaimed in the cache.

We free the WT_REF.page_del structure in 6 different places, add a small inline function to do
the work.