Allow VS Toolchain to be passed a script parameter
To make the script more flexible and future proof, added an
additional parameter that allows the user to specify a vs
toolchain path. This making it easier to move between distro
versions in the evergreen configuration. Also pass this toolchain
file from the evergreen config to better document the toolchain
we expect to find on a Windows VS2017 instance.