WT-7649 Skip timestamp assert during recovery
We don't expect recovery to be using timestamps when applying
commits. If a table is configured with write timestamp asserts,
being "assert=(write_timestamp=on),write_timestamp_usage=always)",
the recovery process will fail. To avoid failing/blocking recovery,
updated the timestamp assert to be skipped on a recovery
connection.