WT-7589 Fixes to allow connection reopen after flush_tier (#6637)
Tiered metadata now inherits from file metadata, so it will get version,
checkpoint, etc. and we can remove the ugly copying of configuration from
the file: metadata to the tiered: metadata.

Fixed block_tiered_load - the block manager's notion of the current objectid
is set explicitly, it is not derived from the root checkpoint.

Made tiered_close and (new function) tiered_discard behave like btree_close
and btree_discard, so they can be called in analogous places.

For a newly created .wtobj local file, close the btree immediately. Otherwise
we have two Btrees that are reading/writing the same file.

When walking metadata looking for btrees, include tiered files, else they
won't be checkpointed.

Recovery now considers tiered files.

When creating config for a tiered object, allocate a new unique Btree id,
and give it a version number.

Set up the local entry in the tiers array when reopening.

Strip out checkpoint when we're creating the (dummy) local file.  Otherwise
it gets the checkpoint from the tiered btree, which definitely does not apply.

Enabled portion of test_tiered02 that does reopens, extended the test as far
as it can go before breaking.

Set the objectid in the checkpoint when writing the extent list.
Set the block's objectid when switching objects.

Expand test_tiered02.py as more things are working.