WT-7231 Evergreen jobs for compiling WiredTiger using CMAKE (#6626)
This PR adds evergreen support for building WiredTiger with CMake. 
It also adds build variants that run the compilation & unit testing tasks
with our CMake builds. The main objective of these new tests are to ensure 
the CMake builds are always working as expected and can correctly 
configure+compile the WiredTiger library and its associated binaries.
The tests under the CMake variants are only a small subset of what is currently 
being tested. This is done to avoid unnecessary duplicate test runs.