WT-7553 cache stuck when fast-delete pages cannot be evicted or rewritten (#6638)
* Loosen the restrictions around evicting fast-truncate pages to avoid cache-stuck failures.

Add a WT_PAGE_DELETED.resolved field that's non-zero once the truncate has been committed or
aborted (but says nothing about which resolution happened or data visibility).

When instantiating a fast-truncate page, use the WT_PAGE_DELETED.resolved field to determine
if we need to create the WT_PAGE_DELETED.update_list array of update references. This doesn't
directly help with evicting fast-truncate pages, but it avoids allocating an additional 8B of
memory for every page entry.

If failing to instantiate a fast-truncate page, don't increment the page's memory footprint.

Don't require global visibility before attempting to evict fast-truncate pages, instead permit
eviction to be attempted as soon as the fast-truncate resolves. This allows fully instantiated
pages to be reconciled and evicted and their memory reclaimed in the cache.

We free the WT_REF.page_del structure in 6 different places, add a small inline function to do
the work.

* Fix an uninitialized variable use in an error path.

* Fix a comment.

* Restore the "can this page be evicted" test on WT_SYNC_DISCARD.

* WT-7579 Disable prefix testing in compatibility test due to snapshot isolation search mismatch (#6641)

* WT-7228 Do not call next if we don't find the key in the history store (#6640)

* WT-7320 Add the ability to check database size to the test framework (#6630)

* WT-7625 Updating throttle configuration to be more user friendly (#6642)

* WT-7588 Make object id numbers 32 bits. (#6636)

* WT-7588 Make object id numbers 32 bits.

* Remove another cast.

* WT-7577 Add sync configuration to flush_tier. (#6606)

* WT-7577 Add sync configuration to flush_tier.

* Typo.

* Run s_all

* Reword to keep s_string happy.

* Add flush state management.

* Remove debugging.

* Remove assert

* The flush condition variable must be destroyed last.

* Simplify - get rid of flush generation number and flush state structure.

* Have the worker thread signal the condition.

* Make state a simple atomic counter, 32 bits. Review comments.

* WT-7594 Use key_consistent mode on format TS runs (#6646)

Updated test/format to configure the table with
'write_timestamp_usage=key_consistent' when doing timestamp runs.
Using 'write_timestamp_usage=always' is incompatible when trying
to use a test/format produced database with other wt utilities
and applications that are not timestamp aware. Resulting in
assertion failures.

* WT-7651 Add synchronization for flush_tier calls. (#6644)

* WT-7651 Add synchronization for flush_tier calls.

* Have lock_wait apply to flush tier schema lock too.

* WT-7656 Destroy condition variable after threads join. (#6650)

* WT-7589 Fixes to allow connection reopen after flush_tier (#6637)

Tiered metadata now inherits from file metadata, so it will get version,
checkpoint, etc. and we can remove the ugly copying of configuration from
the file: metadata to the tiered: metadata.

Fixed block_tiered_load - the block manager's notion of the current objectid
is set explicitly, it is not derived from the root checkpoint.

Made tiered_close and (new function) tiered_discard behave like btree_close
and btree_discard, so they can be called in analogous places.

For a newly created .wtobj local file, close the btree immediately. Otherwise
we have two Btrees that are reading/writing the same file.

When walking metadata looking for btrees, include tiered files, else they
won't be checkpointed.

Recovery now considers tiered files.

When creating config for a tiered object, allocate a new unique Btree id,
and give it a version number.

Set up the local entry in the tiers array when reopening.

Strip out checkpoint when we're creating the (dummy) local file.  Otherwise
it gets the checkpoint from the tiered btree, which definitely does not apply.

Enabled portion of test_tiered02 that does reopens, extended the test as far
as it can go before breaking.

Set the objectid in the checkpoint when writing the extent list.
Set the block's objectid when switching objects.

Expand test_tiered02.py as more things are working.

* Don't look inside the WT_PAGE_DELETED structure without holding the WT_REF locked.

* Fix a few comments, minor cleanups.

* Instead of resolving the fast-truncate transaction on abort, discard the fast-truncate structures,
it's the same result and cleans up memory sooner.

* Don't count the update list as part of the page's footprint, we don't decrement the cache
footprint when freeing it.

* It's not safe to proceed to eviction until the transaction has completed commit, we can switch
to abort after "resolving" a particular truncate operation.

* Fix some comments.

* Split transaction commit processing into two parts, separating out fast-truncate processing.
We have to mark the pages ready for eviction, and that can't be done until the commit can no
longer fail.

* Remove a left-over set of the WT_PAGE_DELETED.resolved flag, that falg has to be set in commit
itself.

Fix a comment.

* When checking if a fast-truncate page can be evicted, we have to acquire the WT_REF lock, but
I think it's possible another thread of control has already locked the WT_REF and is evicting
it, and it will remove the WT_PAGE_DELETED structure as part of that eviction, and the check
could race with that thread. Move the check from __wt_page_can_evict() to __wt_evict_review(),
after eviction has locked the WT_REF.

* The WT_REF.WT_PAGE_DELETED structure must exist, eviction can't have proceeded and removed it
until after the transaction commits.

* Change __wt_overwrite_and_free() and __wt_overwrite_and_free_len() to support calling
with a NULL pointer, similarly to __wt_free().

* Pull the list of WT_UPDATEs from the WT_PAGE_DELETED structure, and union it with the
WT_PAGE_DELETED reference in the WT_REF. This allows a lock-free check of whether or
not an instantiated, fast-truncate page can be evicted.

* Redo naming for WT_REF.fast-truncate structures.

* Rework a comment based on review comments.

* Don't call __wt_overwrite_and_free() unless we know the fast-truncate object size.

* Simplify timestamp assignment in a couple of places -- semantic sugar, no difference in
behavior.

* Add a bigger comment around the WT_REF.{del,update} declarations.

* We don't need an explicit test for fast-truncate pages in __evict_review(), that function calls
__wt_page_can_evict() a little later which will handle the fast-truncate case.

Co-authored-by: Monica Ng <30430618+mm-ng@users.noreply.github.com>
Co-authored-by: Chenhao Qu <chenhao.qu@mongodb.com>
Co-authored-by: Alex Cameron <alex.cameron@10gen.com>
Co-authored-by: Luke Pearson <luke.pearson@mongodb.com>
Co-authored-by: sueloverso <sue@mongodb.com>
Co-authored-by: Alison Felizzi <76922497+alisonfel@users.noreply.github.com>
Co-authored-by: Don Anderson <dda@mongodb.com>