Rework debugging asserts that we're not discarding an internal page with an active page-split generation to consistently check for handle dead and exclusive, exclusive handles cannot be in danger of another thread of control accessing a page-index field. (The bug this is fixing is __wt_page_can_evict() could return the page was evicatable because the handle was exclusive, but the assert in __wt_evict() didn't check exclusivity and so asserted that the page could not be evicted.
Rework rollback-to-stable to protect the page-index only where it's needed, when reviewing
the internal pages for fast-delete leaf pages. This isn't a performance or correctness issue,
it's just clarifying when page-generations are interesting and when they're not (tree walk
handles its own page-generations issues, there's no point in RTS doing it as well).