ext/storage_sources/local_store: fix strict build failure with gcc
On gcc8 and also gcc10, the value range analysis is primitive or maybe
missing entirely and it can't tell that (x % 1000) * 1000 (for
unsigned x) is less than INT_MAX. So if -Wsign-conversion is enabled,
as we do with strict builds, it spuriously complains about unsigned to
signed conversion possibly changing the value. Add explicit cast to
suseconds_t (the type of ->tv_usec in struct timeval according to
POSIX) to silence this.