test/csuite/random_directio: honor configured O_DIRECT blocksize
The code here was using the block size reported by stat.st_blksize as
the alignment requirement for O_DIRECT I/O. This is not correct; it
should be using the O_DIRECT blocksize from the build configuration.

- Distinguish the block alignment from the block size; use
WT_BUFFER_ALIGNMENT_DEFAULT as the alignment, and use the block size
from stat to replace this only if WT_BUFFER_ALIGNMENT_DEFAULT is 0.

- Adjust the assertions about the space in the copy buffer; there
should be space for one block, plus the block alignment so the block
can begin at any offset modulo the block alignment.

- Align the buffer start by the block alignment. At EOF, align the
size up by only the alignment (that's all that's necessary).

- Make the buffer larger. st_blksize typically reports the filesystem
block size, which is mostly 4K on Linux, but on BSD because FFS
supports fragments it's often larger. Most of my volumes are 16K or
32K, and I know there are 64K volumes out there. A 20K buffer isn't
big enough, so make it 64K as a starting point.