Disable Wpoison-system-directories on Darwin
If we are compiling with the strict setting and are not
cross-compiling, we can safely disable the
'Wpoison-system-directories' diagnostic on Clang. This diagnostic
is incompatible with the strict setting when including external
libraries that are not in the default linker path. This issue only
surfaces on the MacOS vendored Clang, where the the
'Wpoison-system-directories' is default enabled.