WT-7748 Fix CMake library probes for libs not on the default linker path (#6720)
* cmake: fix library probes for libs not on the default linker path

On a lot of systems external libraries are in places like /usr/local
or /usr/pkg or /opt that aren't in the default linker path. Use
find_library() instead of check_library_exists() so they can be found
according to user-supplied cmake paths.

When linking with the library, use the path found instead of just the
library name. This is what the cmake manual recommends, rather than
adding to the library path and generating -L linker options.

Also add a HEADER option to config_lib(), use it to look for the
library's header files, and add the resulting dir to the include
path. Pass a header name to uses of config_lib() where this makes
sense.

And, print what's happening in config_lib() since otherwise it's
completely silent.

* helpers.cmake: typo in error message

* Remove the old code for linking the extensions.

(instead of commenting it out)

* cmake: use the explicit library paths for zlib, like zstd/snappy/lz4.

Accidentally left out of the original commit.

* Disable Wpoison-system-directories on Darwin

If we are compiling with the strict setting and are not
cross-compiling, we can safely disable the
'Wpoison-system-directories' diagnostic on Clang. This diagnostic
is incompatible with the strict setting when including external
libraries that are not in the default linker path. This issue only
surfaces on the MacOS vendored Clang, where the the
'Wpoison-system-directories' is default enabled.

Co-authored-by: Alison Felizzi <alison.felizzi@mongodb.com>