Reworked Doxygen CTest description
To avoid confusion, fixed inconsistencies regarding the argument
position of  `-j $(nproc)`, removed the angle brackets in the -R
case (as the command does not require it) & fixed a cut & paste
error.
Also updated the structure of the description to focus on running
all the tests first and then refining down to running specific
tests.