WT-7765 Fix signed integer overflow in intpack-test3 (#6746)
The C standard specifies that overflow in signed integers has undefined behaviour.

This test code tests a range of values around a signed start value. If the test value is (or close to) the maximum value for that int type (in this case INT64_MAX) then the range will overflow, and this is what was being detected by the UBSAN test.

The fix is to change the test ranges where required so that overflows are no longer possible.