WT-7770 Fix issue linking TCMalloc in CMake (#6743)
* ENABLE_TCMALLOC to define HAVE_LIBTCMALLOC

Updated the CMake configuration header to use the ENABLE_TCMALLOC
option to define HAVE_LIBTCMALLOC. Currently autoconf and
mongodb SCons define HAVE_LIBTCMALLOC in the configuration header
when enabling tcmalloc builds. Hence our sources will include
the tcmalloc headers when HAVE_LIBTCMALLOC is defined.  In CMake,
we want to reserve the CPP defintion HAVE_LIBTCMALLOC to just
check if the library exists, not to necessarily link against
tcmalloc. We ideally want to use the ENABLE_TCMALLOC option to
actually create a tcmalloc build.

* Link tcmalloc library when ENABLE_TCMALLOC is set

Added the missing logic to actually link the tcmalloc library when
the user has specified tcmalloc builds.

* Remove unused config_lib code

The CMake config_lib function no longer depends on the
check_library_exists macro and thus does not need to pass in a
function parameter. We can safetly remove this code.