WT-7781 Check for CMake ext path in findExtension (#6750)
The CMake build exports the extension libraries to a different
build path than libtool/autoconf. This results in any test
depending on the extension libraries/modules getting skipped when
using CMake builds. To fix this, when searching for extension
libraries in the python unit test framework, we want to also search
for the location CMake will store the library/module binaries.