WT-7746 Re-sync directory create_test_executable
The CMake helper 'create_test_executable' supports an option to
sync a directory over to the build directory:
'ADDITIONAL_DIRECTORIES'. When creating a test executable we may
want to sync the contents of a directory (such as a set of
configs). Additionally we want this directory to be re-synced on
subsequent builds. CMake didn't detect and sync over new files
and changes added to the source directory. The commit overhauls
the 'ADDITIONAL_DIRECTORIES' feature to ensure new files & changes
get synced over if present. The downside of this change is this
process needs to be added to the 'all' target since CMake can't
automatically detect if new files have been added (which is
reasonable).