Fixed make_directory cmd in define_test_variants
Updated the 'make_directory' step in the 'define_test_variants'
to avoid being re-run on subsequent builds. The command was being
run when the directory already existed. Though this didn't really
have an impact on the test, it adds unnecessary commands to the
'all' target that may not need to be re-run, particularly on clean
builds.