WT-7784 Enable RTS to use checkpoint snapshot on timestamp tables (#6752)
Usually, for timestamped tables, the commit/durable timestamp will be
always greater than the stable timestamp and there is no need of
checkpoint snapshot to abort unstable updates. Allowing it to use
checkpoint snapshot shouldn't be a problem unless the commit/durable
timestamp doesn't adhere to the condition that it must be less than the
stable timestamp.