WT-7831 Clear transaction ids from previous run when repack the cell
To support transaction id based rollback to stable, WT didn't reset
the transaction ids stored on the pages by not incrementing the
base_write_gen of each btree upon page access until rollback to stable
operates on that page. Once RTS operates on that page and reset the
remaining transaction ids that are present on the page that are stable.