WT-7746 Improve directory syncing with CMake helper 'create_test_executable' (#6762)
* Re-sync directory create_test_executable

The CMake helper 'create_test_executable' supports an option to
sync a directory over to the build directory:
'ADDITIONAL_DIRECTORIES'. When creating a test executable we may
want to sync the contents of a directory (such as a set of
configs). Additionally we want this directory to be re-synced on
subsequent builds. CMake didn't detect and sync over new files
and changes added to the source directory. The commit overhauls
the 'ADDITIONAL_DIRECTORIES' feature to ensure new files & changes
get synced over if present. The downside of this change is this
process needs to be added to the 'all' target since CMake can't
automatically detect if new files have been added (which is
reasonable).

* Fixed make_directory cmd in define_test_variants

Updated the 'make_directory' step in the 'define_test_variants'
to avoid being re-run on subsequent builds. The command was being
run when the directory already existed. Though this didn't really
have an impact on the test, it adds unnecessary commands to the
'all' target that may not need to be re-run, particularly on clean
builds.

* Remove manual CMake swig lib rename

Removed the post build command renaming the swig library from
'__wiredtiger' to '_wiredtiger'. We can specify the targets output
name via the target propety 'OUTPUT_NAME'. Letting CMake do the
renaming for us. This has the added benefit of avoiding swig being
re-run on clean/incremental builds when no changes have been made.
This was happening since the POST_BUILD rename command was
constantly dirtying the swig target.