Try again with the comment formatting.
It seems that the header is required to use hanging indent by
function.py, so in order to avoid the rest of the comments after being
reformatted with hanging indent by clang-format, move them inside the
function body.

This is maybe not optimal but it at least isn't visually revolting and
doesn't break the tree :-|

Also, add sodium_encrypt.c to dist/extlist so that all the checks are
run on it.